generate_data/requirements.txt
Faker
google-cloud-pubsub
apache_beam[gcp]
